Allow specifying --trait-col for association results
If association results already has a trait column, we can use it instead
of directly specifying the --trait. This is useful in the scenario where
you lump together results for multiple traits into a single file, to
pick lead SNPs at each locus by which trait has the strongest
association.